feat: convert to ESM + update all dependencies
BREAKING CHANGE: We now require node 20 or higher

This is an internal only package so a good candidate for early updates to be ESM compatible, makes it easier to update packages and maintain moving forward. Has no impact on consumption of this module as we exclusively use it via the CLI in the electron repo.